+++ Via.h
@@ -32,7 +32,7 @@ public:
 	void setDatosDemograficos(DatosDemograficos* dD);
 	int alg7(int limInf, int limSup);
 	void alg8(ListaPI<Estudios*>*& lEstud);
-	void alg9(int& nHombres, int& nMujeres);
+	void alg9(float& nHombres, float& nMujeres);
 
 	void mostrar();
